Struct ink_storage::alloc::allocator::DynamicAllocator

pub struct DynamicAllocator {
    allocations: BitStash,
}

The dynamic allocator.

Manages dynamic storage allocations in a very efficient and economic way.

Fields
allocations: BitStash
Implementations

impl DynamicAllocator

pub fn alloc(&mut self) -> DynamicAllocation

Returns a new dynamic storage allocation.

Panics

If the dynamic allocator ran out of free dynamic allocations.

pub fn free(&mut self, allocation: DynamicAllocation)

Frees the given dynamic storage allocation.

This makes the given dynamic storage allocation available again for new dynamic storage allocations.

Panics

Panics if the given dynamic allocation is invalid. A dynamic allocation is invalid if it is not represented as occupied in the free list.