Function ink_storage::alloc::initialize

pub fn initialize(phase: ContractPhase)

Tells the global dynamic storage allocator instance how it shall initialize.

Note

Normally users of ink! do not have to call this function directly as it is automatically being use in the correct order and way by the generated code.

The phase parameter describes for which execution phase the dynamic storage allocator needs to be initialized since this is different in contract instantiations and calls.

This has to be issued before the first interaction with the global allocator.

The actual instantiation will happen only upon the first interaction with the global allocator, e.g. using the alloc or free calls. Until then, it remains uninitialized.

If this function is not called before the first global allocator interaction then the default initialization scheme is for contract instantiation. However, this behavior might change and must not be relied upon.