Struct ink_storage::alloc::DynamicAllocation

pub struct DynamicAllocation(_);

A unique dynamic allocation.

This can refer to a dynamically allocated storage cell. It has been created by a dynamic storage allocator. The initiator of the allocation has to make sure to deallocate this dynamic allocation again using the same dynamic allocator if it is no longer in use.

Note

Normally instances of this type are not used directly and instead a storage::Box is used instead.

Implementations

impl DynamicAllocation

pub fn key(self) -> Key

Returns the storage key associated with this dynamic allocation.