Module ink_storage::collections

High-level collections used to manage storage entities in the persisted contract storage.

Users should generally use these collections in their contracts directly or as building blocks for their collections and algorithms.

Modules

binary_heap
A priority queue implemented with a binary heap.

bitstash
Storage bit stash data structure and utilities.

bitvec
Storage bit vector data structure and utilities.

hashmap
A storage hash map that allows to associate keys with values.

smallvec
A small storage vector that allows to store a limited amount of elements.

stash
A storage stash allowing to store indexed elements efficiently.

vec
A storage vector used to store elements in a contiguous sequenced order.

Structs

BinaryHeap
A priority queue implemented with a binary heap.

BitStash
A stash for bits operating on the contract storage.

Bitvec
A storage bit vector.

HashMap
A hash map operating on the contract storage.

SmallVec
A contiguous growable array type.

Stash
A stash data structure operating on contract storage.

Vec
A contiguous growable array type, written Vec<T> but pronounced “vector”.