Module ink_storage::collections::binary_heap

A priority queue implemented with a binary heap.

Insertion and popping the largest element have O(log(n)) complexity. Checking the largest element is O(1).

Structs

BinaryHeap

A priority queue implemented with a binary heap.

Iter

An iterator over shared references to the elements of the BinaryHeap.

PeekMut

Structure wrapping a mutable reference to the greatest item on a BinaryHeap.

Reverse

Wrapper for core::cmp::Reverse for using a BinaryHeap as a min-heap.