Struct ink_storage::collections::bitstash::counts::CountFree

pub struct CountFree {
    counts: [u8; 32],
    full: FullMask,
}

Stores the number of set bits for each 256-bits block in a compact u8.

Fields
counts: [u8; 32]

Set bits per 256-bit chunk.

full: FullMask

Since with u8 can only count up to 255 but there might be the need to count up to 256 bits for 256-bit chunks we need to store one extra bit per counter to determine filled chunks.

Implementations

impl CountFree

pub fn position_first_zero(&self) -> Option<u8>

Returns the position of the first free u8 in the free counts.

Returns None if all counts are 0xFF.

pub fn inc(&mut self, index: usize)

Increases the number of set bits for the given index.

Panics

If the given index is out of bounds.

If the increment would cause an overflow.

pub fn dec(&mut self, index: u8) -> u8

Decreases the number of set bits for the given index.

Returns the new number of set bits.

Panics

If the given index is out of bounds.

If the decrement would cause an overflow.