Struct ink_storage::collections::bitvec::bits256::Bits256

pub struct Bits256 {
    bits: [u64; 4],
}

A chunk of 256 bits.

Fields
bits: [u64; 4]
Implementations

impl Bits256

pub(in collections::bitvec::bits256) fn bits_at(
&self,
index: u8
) -> (&u64, u8)

pub(in collections::bitvec::bits256) fn bits_at_mut(
&mut self,
index: u8
) -> (&mut u64, u8)

pub(in collections::bitvec) fn iter(
&self,
len: u16
) -> Iter<'_>

Yields the first len bits of the pack of 256 bits.

pub(in collections::bitvec) fn iter_mut(
&mut self,
len: u16
) -> IterMut<'_>

Yields mutable accessors to the first len bits of the pack of 256 bits.

pub fn get(&self, at: u8) -> bool

Returns the bit value for the bit at the given index.

pub(in collections::bitvec) fn set_to(
&mut self,
at: u8,
new_value: bool
)

Sets the bit value for the bit at the given index to the given value.

pub(in collections::bitvec) fn flip(
&mut self,
at: u8
)

Flips the bit value for the bit at the given index.

pub(in collections::bitvec) fn set(
&mut self,
at: u8
)

Sets the bit value for the bit at the given index to 1 (true).

pub(in collections::bitvec) fn reset(
&mut self,
at: u8
)

Sets the bit value for the bit at the given index to 0 (false).

pub(in collections::bitvec::bits256) fn op_at_with<F>(
&mut self,
at: u8,
rhs: bool,
op: F
) where
F: FnOnce(&mut u64, u64),

pub(in collections::bitvec) fn and(
&mut self,
at: u8,
rhs: bool
)

Computes bitwise AND for the bit at the given index and rhs.

pub(in collections::bitvec) fn or(
&mut self,
at: u8,
rhs: bool
)

Computes bitwise OR for the bit at the given index and rhs.

pub(in collections::bitvec) fn xor(
&mut self,
at: u8,
rhs: bool
)

Computes bitwise XOR for the bit at the given index and rhs.

pub fn position_first_zero(&self) -> Option<u8>

Returns the position of the first zero bit if any.