Struct ink_storage::collections::bitvec::BitRefMut

pub struct BitRefMut<'a> { /* fields omitted */ }

A mutable bit access for operating on a single bit within a 256-bit pack.

Implementations

impl<'a> BitRefMut<'a>

pub fn get(&self) -> bool

Returns the value of the indexed bit.

Note

If 0: returns false

If 1: returns true

pub fn set_to(&mut self, new_value: bool)

Sets the value of the indexed bit to the given new value.

pub fn set(&mut self)

Sets the indexed bit to 1 (true).

pub fn reset(&mut self)

Resets the indexed bit to 0 (false).

pub fn flip(&mut self)

Flips the indexed bit.

pub fn xor(&mut self, rhs: bool)

Computes bitwise XOR for the indexed bit and rhs.

pub fn and(&mut self, rhs: bool)

Computes bitwise AND for the indexed bit and rhs.

pub fn or(&mut self, rhs: bool)

Computes bitwise OR for the indexed bit and rhs.