Struct ink_storage::collections::bitvec::Bitvec

pub struct Bitvec { /* fields omitted */ }

A storage bit vector.

Note

Organizes its bits in chunks of 256 bits. Allows to push, pop, inspect and manipulate the underlying bits.

Implementations

impl Bitvec

pub fn new() -> Self

Creates a new empty bit vector.

pub fn len(&self) -> u32

Returns the length of the bit vector in bits.

pub fn is_empty(&self) -> bool

Returns true if the bit vector is empty.

pub fn capacity(&self) -> u64

Returns the capacity of the bit vector in bits.

Note

Returns a u64 since it is always greater than or equal to self.len() which itself returns a u32.

pub fn bits(&self) -> BitsIter<'_>

Returns an iterator over the bits of the storage bit vector.

pub fn bits_mut(&mut self) -> BitsIterMut<'_>

Returns an iterator over the mutable bits of the storage bit vector.

pub fn get(&self, at: u32) -> Option<bool>

Returns the value of the bit at the given index if any.

pub fn get_mut(&mut self, at: u32) -> Option<BitRefMut<'_>>

Returns a mutable bit access to the bit at the given index if any.

pub fn get_chunk(&self, at: u32) -> Option<ChunkRef<&Bits256>>

Returns a shared reference to the 256-bit chunk for the bit at the given index.

pub fn get_chunk_mut(&mut self, at: u32) -> Option<ChunkRef<&mut Bits256>>

Returns an exclusive reference to the 256-bit chunk for the bit at the given index.

pub fn first(&self) -> Option<bool>

Returns the first bit of the bit vector.

Note

Returns None if the bit vector is empty.

pub fn first_mut(&mut self) -> Option<BitRefMut<'_>>

Returns a mutable bit access to the first bit of the bit vector.

Note

Returns None if the bit vector is empty.

pub fn last(&self) -> Option<bool>

Returns the last bit of the bit vector.

Note

Returns None if the bit vector is empty.

pub fn last_mut(&mut self) -> Option<BitRefMut<'_>>

Returns a mutable bit access to the last bit of the bit vector.

Note

Returns None if the bit vector is empty.

pub fn push(&mut self, value: bool)

Pushes the given value onto the bit vector.

Note

This increases the length of the bit vector.

Panics

If the storage bit vector reached its maximum capacity.

pub fn pop(&mut self) -> Option<bool>

Pops the last bit from the bit vector.

Returns the popped bit as bool.

Note

This reduces the length of the bit vector by one.