Struct ink_storage::collections::hashmap::HashMap

pub struct HashMap<K, V, H = Blake2x256> where
    K: Ord + Clone + PackedLayout,
    V: PackedLayout,
    H: CryptoHash,
    Key: From<<H as HashOutput>::Type>,  { /* fields omitted */ }

A hash map operating on the contract storage.

Stores a mapping between keys and values.

Note

Unlike Rust's standard HashMap that uses the core::hash::Hash trait in order to hash its keys the storage hash map uses the [scale::Encode] encoding in order to hash its keys using a built-in cryptographic hash function provided by the chain runtime.

The main difference between the lower-level LazyHashMap and the storage::HashMap is that the latter is aware of its associated keys and values and operates on those instances directly as opposed to Option instances of them. Also it provides a more high-level and user focused API.

Users should generally prefer using this storage hash map over the low-level LazyHashMap for direct usage in their smart contracts.

Implementations

impl<K, V, H> HashMap<K, V, H> where
K: Ord + Clone + PackedLayout,
V: PackedLayout,
H: CryptoHash,
Key: From<<H as HashOutput>::Type>,

pub fn new() -> Self

Creates a new empty storage hash map.

pub fn len(&self) -> u32

Returns the number of key-value pairs stored in the hash map.

pub fn is_empty(&self) -> bool

Returns true if the hash map is empty.

pub fn iter(&self) -> Iter<'_, K, V, H>

Returns an iterator yielding shared references to all key/value pairs of the hash map.

Note

Avoid unbounded iteration over big storage hash maps.

Prefer using methods like Iterator::take in order to limit the number of yielded elements.

pub fn iter_mut(&mut self) -> IterMut<'_, K, V, H>

Returns an iterator yielding exclusive references to all key/value pairs of the hash map.

Note

Avoid unbounded iteration over big storage hash maps.

Prefer using methods like Iterator::take in order to limit the number of yielded elements.

pub fn values(&self) -> Values<'_, K, V, H>

Returns an iterator yielding shared references to all values of the hash map.

Note

Avoid unbounded iteration over big storage hash maps.

Prefer using methods like Iterator::take in order to limit the number of yielded elements.

pub fn values_mut(&mut self) -> ValuesMut<'_, K, V, H>

Returns an iterator yielding shared references to all values of the hash map.

Note

Avoid unbounded iteration over big storage hash maps.

Prefer using methods like Iterator::take in order to limit the number of yielded elements.

pub fn keys(&self) -> Keys<'_, K>

Returns an iterator yielding shared references to all keys of the hash map.

Note

Avoid unbounded iteration over big storage hash maps.

Prefer using methods like Iterator::take in order to limit the number of yielded elements.

impl<K, V, H> HashMap<K, V, H> where
K: Ord + Eq + Clone + PackedLayout,
V: PackedLayout,
H: CryptoHash,
Key: From<<H as HashOutput>::Type>,

pub fn insert(&mut self, key: K, new_value: V) -> Option<V>

Inserts a key-value pair into the map.

Returns the previous value associated with the same key if any. If the map did not have this key present, None is returned.

Note

If the map did have this key present, the value is updated, and the old value is returned. The key is not updated, though; this matters for types that can be == without being identical.

pub fn take<Q>(&mut self, key: &Q) -> Option<V> where
K: Borrow<Q>,
Q: Ord + Encode + ToOwned<Owned = K>,

Removes the key/value pair from the map associated with the given key.

Returns the removed value if any.

Note

The key may be any borrowed form of the map's key type, but Hash and Eq on the borrowed form must match those for the key type.

pub fn get<Q>(&self, key: &Q) -> Option<&V> where
K: Borrow<Q>,
Q: Ord + Encode + ToOwned<Owned = K>,

Returns a shared reference to the value corresponding to the key.

The key may be any borrowed form of the map's key type, but Hash and Eq on the borrowed form must match those for the key type.

pub fn get_mut<Q>(&mut self, key: &Q) -> Option<&mut V> where
K: Borrow<Q>,
Q: Ord + Encode + ToOwned<Owned = K>,

Returns a mutable reference to the value corresponding to the key.

The key may be any borrowed form of the map's key type, but Hash and Eq on the borrowed form must match those for the key type.

pub fn contains_key<Q>(&self, key: &Q) -> bool where
K: Borrow<Q>,
Q: Ord + PartialEq<K> + Eq + Encode + ToOwned<Owned = K>,

Returns true if there is an entry corresponding to the key in the map.

pub fn defrag(&mut self, max_iterations: Option<u32>) -> u32

Defragments storage used by the storage hash map.

Returns the number of storage cells freed this way.

A max_iterations parameter of None means that there is no limit to the number of iterations performed. This is generally not advised.

Note

This frees storage that is held but not necessary for the hash map to hold. This operation might be expensive, especially for big max_iteration parameters. The max_iterations parameter can be used to limit the expensiveness for this operation and instead free up storage incrementally.

pub fn entry(&mut self, key: K) -> Entry<'_, K, V>

Gets the given key's corresponding entry in the map for in-place manipulation.