Struct ink_storage::collections::smallvec::SmallVec

pub struct SmallVec<T, const N: usize> where
    T: PackedLayout,  { /* fields omitted */ }

A contiguous growable array type.

Note

The storage::SmallVec has a very similar API compared to a storage::Vec. The major difference between both data structures is that the SmallVec can only contain up to a fixed amount of elements given by N whereas the Vec can contain up to 2^32 elements which is the maximum for 32-bit Wasm targets.

The performance characteristics may be different from Rust's Vec due to the internal differences.

Allows to store up to N elements.

Implementations

impl<T, const N: usize> SmallVec<T, N> where
T: PackedLayout,

pub fn new() -> Self

Creates a new empty vector.

pub fn capacity(&self) -> u32

Returns the capacity of the small vector.

pub fn len(&self) -> u32

Returns the number of elements in the vector, also referred to as its 'length'.

pub fn is_empty(&self) -> bool

Returns true if the vector contains no elements.

impl<T, const N: usize> SmallVec<T, N> where
T: PackedLayout,

pub fn iter(&self) -> Iter<'_, T, N>

Returns an iterator yielding shared references to all elements.

Note

Avoid unbounded iteration over big storage vectors.

Prefer using methods like Iterator::take in order to limit the number of yielded elements.

pub fn iter_mut(&mut self) -> IterMut<'_, T, N>

Returns an iterator yielding exclusive references to all elements.

Note

Avoid unbounded iteration over big storage vectors.

Prefer using methods like Iterator::take in order to limit the number of yielded elements.

pub fn first(&self) -> Option<&T>

Returns a shared reference to the first element if any.

pub fn last(&self) -> Option<&T>

Returns a shared reference to the last element if any.

pub fn get(&self, index: u32) -> Option<&T>

Returns a shared reference to the indexed element.

Returns None if index is out of bounds.

impl<T, const N: usize> SmallVec<T, N> where
T: PackedLayout,

pub fn push(&mut self, value: T)

Appends an element to the back of the vector.

impl<T, const N: usize> SmallVec<T, N> where
T: PackedLayout,

pub fn pop(&mut self) -> Option<T>

Pops the last element from the vector and returns it. Returns None if the vector is empty.

pub fn pop_drop(&mut self) -> Option<()>

Pops the last element from the vector and immediately drops it.

Returns Some(()) if an element has been removed and None otherwise.

Note

This operation is a bit more efficient than SmallVec::pop since it avoids reading from contract storage in some use cases.

pub fn first_mut(&mut self) -> Option<&mut T>

Returns an exclusive reference to the first element if any.

pub fn last_mut(&mut self) -> Option<&mut T>

Returns an exclusive reference to the last element if any.

pub fn get_mut(&mut self, index: u32) -> Option<&mut T>

Returns an exclusive reference to the indexed element.

Returns None if index is out of bounds.

pub fn swap(&mut self, a: u32, b: u32)

Swaps the elements at the given indices.

Panics

If one or both indices are out of bounds.

pub fn swap_remove(&mut self, n: u32) -> Option<T>

Removes the indexed element from the vector and returns it.

The last element of the vector is put into the indexed slot. Returns None and does not mutate the vector if the index is out of bounds.

Note

This operation does not preserve ordering but is constant time.

pub fn swap_remove_drop(&mut self, n: u32) -> Option<()>

Removes the indexed element from the vector.

The last element of the vector is put into the indexed slot. Returns Some(()) if an element has been removed and None otherwise.

Note

This operation should be preferred over Vec::swap_remove if there is no need to return the removed element since it avoids a contract storage read for some use cases.