Module ink_storage::collections::stash

A storage stash allowing to store indexed elements efficiently.

Structs

Iter

An iterator over shared references to the elements of a storage stash.

IterMut

An iterator over exclusive references to the elements of a storage stash.

Stash

A stash data structure operating on contract storage.

VacantEntry

A vacant entry with previous and next vacant indices.

Enums

Entry

An entry within the stash.