Enum ink_storage::collections::stash::Entry

pub enum Entry<T> {
    Vacant(VacantEntry),
    Occupied(T),
}

An entry within the stash.

The vacant entries within a storage stash form a doubly linked list of vacant entries that is used to quickly re-use their vacant storage.

Variants

Vacant(VacantEntry)

A vacant entry that holds the index to the next and previous vacant entry.

Occupied(T)

An occupied entry that hold the value.

Implementations

impl<T> Entry<T>

pub fn is_occupied(&self) -> bool

Returns true if the entry is occupied.

pub fn is_vacant(&self) -> bool

Returns true if the entry is vacant.