Struct ink_storage::collections::stash::Header

pub(in collections::stash) struct Header {
    last_vacant: u32,
    len: u32,
    len_entries: u32,
}

Stores general commonly required information about the storage stash.

Fields
last_vacant: u32

The latest vacant index.

If all entries are occupied:

Points to the entry at index self.len.

If some entries are vacant:

Points to the entry that has been vacated most recently.

len: u32

The number of items stored in the stash.

Note

We cannot simply use the underlying length of the vector since it would include vacant slots as well.

len_entries: u32

The number of entries currently managed by the stash.