Struct ink_storage::collections::stash::Stash

pub struct Stash<T> where
    T: PackedLayout,  { /* fields omitted */ }

A stash data structure operating on contract storage.

This allows to store information similar to a vector but in unordered fashion which enables constant time random deletion of elements. This allows for efficient attachment of data to some numeric indices.

Implementations

impl<T> Stash<T> where
T: PackedLayout,

pub fn new() -> Self

Creates a new empty stash.

pub fn len(&self) -> u32

Returns the number of elements stored in the stash.

pub fn is_empty(&self) -> bool

Returns true if the stash contains no elements.

pub fn capacity(&self) -> u32

Returns the number of entries the stash can hold without allocating another storage cell.

Note

This is the total number of occupied and vacant entries of the stash.

pub fn entries_key(&self) -> Option<&Key>

Returns the underlying key to the cells.

Note

This is a low-level utility getter and should normally not be required by users.

pub fn iter(&self) -> Iter<'_, T>

Returns an iterator yielding shared references to all elements of the stash.

Note

Avoid unbounded iteration over big storage stashes. Prefer using methods like Iterator::take in order to limit the number of yielded elements.

pub fn iter_mut(&mut self) -> IterMut<'_, T>

Returns an iterator yielding exclusive references to all elements of the stash.

Note

Avoid unbounded iteration over big storage stashes. Prefer using methods like Iterator::take in order to limit the number of yielded elements.

pub fn entries(&self) -> Entries<'_, T>

Returns an iterator yielding shared references to all entries of the stash.

impl<T> Stash<T> where
T: PackedLayout,

pub fn get(&self, at: u32) -> Option<&T>

Returns a shared reference to the element at the given index.

pub fn get_mut(&mut self, at: u32) -> Option<&mut T>

Returns an exclusive reference to the element at the given index.

impl<T> Stash<T> where
T: PackedLayout,

pub fn put(&mut self, new_value: T) -> u32

Put the element into the stash at the next vacant position.

Returns the stash index that the element was put into.

pub fn take(&mut self, at: u32) -> Option<T>

Takes the element stored at the given index if any.

pub unsafe fn remove_occupied(&mut self, at: u32) -> Option<()>

Removes the element stored at the given index if any.

This method acts similar to the take API and even still returns an Option. However, it guarantees to make no contract storage reads to the indexed element and will only write to its internal low-level lazy cache that the element at the given index is going to be removed at the end of the contract execution.

Calling this method with an index out of bounds for the returns None and does not remove the element, otherwise it returns Some(()).

Safety

The caller must ensure that at refers to an occupied index. Behavior is unspecified if at refers to a vacant index and could seriously damage the contract storage integrity.

pub fn defrag<C>(&mut self, max_iterations: Option<u32>, callback: C) -> u32 where
C: FnMut(u32, u32, &T),

Defragments the underlying storage to minimize footprint.

Returns the number of storage cells freed this way.

This might invalidate indices stored outside the stash.

Callback

In order to keep those indices up-to-date the caller can provide a callback function that is called for every moved entry with a shared reference to the entries value and the old as well as the new index.

Note

If max_iterations is Some concrete value it is used in order to bound the number of iterations and won’t try to defrag until the stash is optimally compacted.

Users are advised to call this method using Some concrete value to keep gas costs within certain bounds.

The call to the given callback takes place before the reinsertion of the shifted occupied entry.