Struct ink_storage::collections::BinaryHeap

pub struct BinaryHeap<T> where
    T: PackedLayout + Ord,  { /* fields omitted */ }

A priority queue implemented with a binary heap.

Note

The heap is a max-heap by default, i.e. the first element is the largest. Either Reverse or a custom Ord implementation can be used to make BinaryHeap a min-heap. This then makes heap.pop() return the smallest value instead of the largest one.

Implementations

impl<T> BinaryHeap<T> where
T: PackedLayout + Ord,

pub fn new() -> Self

Creates a new empty storage heap.

pub fn len(&self) -> u32

Returns the number of elements in the heap, also referred to as its length.

pub fn is_empty(&self) -> bool

Returns true if the heap contains no elements.

impl<T> BinaryHeap<T> where
T: PackedLayout + Ord,

pub fn iter(&self) -> Iter<'_, T>

Returns an iterator yielding shared references to all elements of the heap.

Note

Avoid unbounded iteration over large heaps. Prefer using methods like Iterator::take in order to limit the number of yielded elements.

pub fn peek(&self) -> Option<&T>

Returns a shared reference to the greatest element of the heap

Returns None if the heap is empty

pub fn peek_mut(&mut self) -> Option<PeekMut<'_, T>>

Returns an exclusive reference to the greatest element of the heap

Returns None if the heap is empty

Note:

If the PeekMut value is leaked, the heap may be in an inconsistent state.

Example

use ink_storage::collections::BinaryHeap;
let mut heap = BinaryHeap::new();
assert!(heap.peek_mut().is_none());

heap.push(1);
heap.push(5);
heap.push(2);
{
    let mut val = heap.peek_mut().unwrap();
    *val = 0;
}
assert_eq!(heap.peek(), Some(&2));

pub fn pop(&mut self) -> Option<T>

Pops greatest element from the heap and returns it

Returns None if the heap is empty

pub fn clear(&mut self)

Removes all elements from this heap.

Note

Use this method to clear the heap instead of e.g. iterative pop(). This method performs significantly better and does not actually read any of the elements (whereas pop() does).

impl<T> BinaryHeap<T> where
T: PackedLayout + Ord,

pub fn push(&mut self, value: T)

Pushes the given element to the binary heap.