Struct ink_storage::collections::BitStash

pub struct BitStash { /* fields omitted */ }

A stash for bits operating on the contract storage.

Allows to efficiently put and take bits and stores the underlying bits in an extremely compressed format.

Implementations

impl BitStash

pub fn new() -> Self

Creates a new storage bit stash.

pub fn get(&self, index: u32) -> Option<bool>

Returns true if the bit at the indexed slot is set (1).

Returns None if the index is out of bounds.

pub fn put(&mut self) -> u32

Puts another set bit into the storage bit stash.

Returns the index to the slot where the set bit has been inserted.

pub fn take(&mut self, index: u32) -> Option<bool>

Takes the bit from the given index and returns it.

Returns true if the indexed bit was set (1). Returns None if the index is out of bounds.

Note

This frees up the indexed slot for putting in another set bit.