Struct ink_storage::collections::Vec

pub struct Vec<T> where
    T: PackedLayout,  { /* fields omitted */ }

A contiguous growable array type, written Vec<T> but pronounced “vector”.

Note

Despite the similarity to Rust’s Vec type this storage Vec has many differences in its internal data layout. While it stores its data in contiguous storage slots this does not mean that the data is actually densely stored in memory.

Also its technical performance characteristics may be different from Rust’s Vec due to the differences stated above.

Allows to store up to 2^32 elements and is guaranteed to not reallocate upon pushing new elements to it.

Implementations

impl<T> Vec<T> where
T: PackedLayout,

pub fn new() -> Self

Creates a new empty storage vector.

pub fn len(&self) -> u32

Returns the number of elements in the vector, also referred to as its length.

pub fn is_empty(&self) -> bool

Returns true if the vector contains no elements.

impl<T> Vec<T> where
T: PackedLayout,

pub fn iter(&self) -> Iter<'_, T>

Returns an iterator yielding shared references to all elements of the vector.

Note

Avoid unbounded iteration over big storage vectors. Prefer using methods like Iterator::take in order to limit the number of yielded elements.

pub fn iter_mut(&mut self) -> IterMut<'_, T>

Returns an iterator yielding exclusive references to all elements of the vector.

Note

Avoid unbounded iteration over big storage vectors. Prefer using methods like Iterator::take in order to limit the number of yielded elements.

pub fn first(&self) -> Option<&T>

Returns a shared reference to the first element if any.

pub fn last(&self) -> Option<&T>

Returns a shared reference to the last element if any.

pub fn get(&self, index: u32) -> Option<&T>

Returns a shared reference to the indexed element.

Returns None if index is out of bounds.

impl<T> Vec<T> where
T: PackedLayout,

pub fn push(&mut self, value: T)

Appends an element to the back of the vector.

pub fn binary_search(&self, x: &T) -> Result<u32, u32> where
T: Ord,

Binary searches this sorted vector for a given element.

If the value is found then Result::Ok is returned, containing the index of the matching element. If there are multiple matches, then any one of the matches could be returned. If the value is not found then Result::Err is returned, containing the index where a matching element could be inserted while maintaining sorted order.

See also binary_search_by, binary_search_by_key.

Examples

Looks up a series of four elements. The first is found, with a uniquely determined position; the second and third are not found; the fourth could match any position in [1, 4].

use ink_storage::Vec as StorageVec;

let s: StorageVec<i32> = [0, 1, 1, 1, 1, 2, 3, 5, 8, 13, 21, 34, 55]
    .into_iter()
    .collect();

assert_eq!(s.binary_search(&13),  Ok(9));
assert_eq!(s.binary_search(&4),   Err(7));
assert_eq!(s.binary_search(&100), Err(13));
let r = s.binary_search(&1);
assert!(match r { Ok(1..=4) => true, _ => false, });

pub fn binary_search_by<'a, F>(&'a self, f: F) -> Result<u32, u32> where
F: FnMut(&'a T) -> Ordering,

Binary searches this sorted vector with a comparator function.

The comparator function should implement an order consistent with the sort order of the underlying vector, returning an order code that indicates whether its argument is Less, Equal or Greater the desired target.

If the value is found then Result::Ok is returned, containing the index of the matching element. If there are multiple matches, then any one of the matches could be returned. If the value is not found then Result::Err is returned, containing the index where a matching element could be inserted while maintaining sorted order.

See also binary_search, binary_search_by_key.

Examples

Looks up a series of four elements. The first is found, with a uniquely determined position; the second and third are not found; the fourth could match any position in [1, 4].

use ink_storage::Vec as StorageVec;

let s: StorageVec<i32> = [0, 1, 1, 1, 1, 2, 3, 5, 8, 13, 21, 34, 55]
    .into_iter()
    .collect();

let seek = 13;
assert_eq!(s.binary_search_by(|probe| probe.cmp(&seek)), Ok(9));
let seek = 4;
assert_eq!(s.binary_search_by(|probe| probe.cmp(&seek)), Err(7));
let seek = 100;
assert_eq!(s.binary_search_by(|probe| probe.cmp(&seek)), Err(13));
let seek = 1;
let r = s.binary_search_by(|probe| probe.cmp(&seek));
assert!(match r { Ok(1..=4) => true, _ => false, });

pub fn binary_search_by_key<'a, B, F>(&'a self, b: &B, f: F) -> Result<u32, u32> where
F: FnMut(&'a T) -> B,
B: Ord,

Binary searches this sorted vector with a key extraction function.

If the value is found then Result::Ok is returned, containing the index of the matching element. If there are multiple matches, then any one of the matches could be returned. If the value is not found then Result::Err is returned, containing the index where a matching element could be inserted while maintaining sorted order.

See also binary_search, binary_search_by.

Examples

Looks up a series of four elements in a vector of pairs sorted by their second elements. The first is found, with a uniquely determined position; the second and third are not found; the fourth could match any position in [1, 4].

use ink_storage::Vec as StorageVec;

let s: StorageVec<(i32, i32)> = [
    (0, 0),
    (2, 1),
    (4, 1),
    (5, 1),
    (3, 1),
    (1, 2),
    (2, 3),
    (4, 5),
    (5, 8),
    (3, 13),
    (1, 21),
    (2, 34),
    (4, 55),
]
.into_iter()
.collect();

assert_eq!(s.binary_search_by_key(&13, |&(a, b)| b),  Ok(9));
assert_eq!(s.binary_search_by_key(&4, |&(a, b)| b),   Err(7));
assert_eq!(s.binary_search_by_key(&100, |&(a, b)| b), Err(13));
let r = s.binary_search_by_key(&1, |&(a, b)| b);
assert!(match r { Ok(1..=4) => true, _ => false, });

impl<T> Vec<T> where
T: PackedLayout,

pub fn pop(&mut self) -> Option<T>

Pops the last element from the vector and returns it. Returns None if the vector is empty.

pub fn pop_drop(&mut self) -> Option<()>

Pops the last element from the vector and immediately drops it.

Returns Some(()) if an element has been removed and None otherwise.

Note

This operation is a bit more efficient than Vec::pop since it avoids reading from contract storage in some use cases.

pub fn first_mut(&mut self) -> Option<&mut T>

Returns an exclusive reference to the first element if any.

pub fn last_mut(&mut self) -> Option<&mut T>

Returns an exclusive reference to the last element if any.

pub fn get_mut(&mut self, index: u32) -> Option<&mut T>

Returns an exclusive reference to the indexed element.

Returns None if index is out of bounds.

pub fn swap(&mut self, a: u32, b: u32)

Swaps the elements at the given indices.

Panics

If one or both indices are out of bounds.

pub fn swap_remove(&mut self, n: u32) -> Option<T>

Removes the indexed element from the vector and returns it.

The last element of the vector is put into the indexed slot. Returns None and does not mutate the vector if the index is out of bounds.

Note

This operation does not preserve ordering but is constant time.

pub fn swap_remove_drop(&mut self, n: u32) -> Option<()>

Removes the indexed element from the vector.

The last element of the vector is put into the indexed slot. Returns Some(()) if an element has been removed and None otherwise.

Note

This operation should be preferred over Vec::swap_remove if there is no need to return the removed element since it avoids a contract storage read for some use cases.

pub fn set(&mut self, index: u32, new_value: T) -> Result<(), IndexOutOfBounds>

Sets the elements at the given index to the new value.

Won’t return the old element back to the caller. Prefer this operation over other method of overriding an element in the storage vector since this is more efficient.

pub fn clear(&mut self)

Removes all elements from this vector.

Note

Use this method to clear the vector instead of e.g. iterative pop(). This method performs significantly better and does not actually read any of the elements (whereas pop() does).