Module ink_storage::collections::vec

A storage vector used to store elements in a contiguous sequenced order.

This is by default the go-to collection for most smart contracts if there are no special requirements to the storage data structure.

Structs

IndexOutOfBounds
The index is out of the bounds of this vector.

Iter
An iterator over shared references to the elements of a storage vector.

IterMut
An iterator over exclusive references to the elements of a storage vector.

Vec
A contiguous growable array type, written Vec<T> but pronounced “vector”.