Module ink_storage::lazy::lazy_hmap

A lazy storage mapping that stores entries under their SCALE encoded key hashes.

Structs

LazyHashMap
A lazy storage mapping that stores entries under their SCALE encoded key hashes.

OccupiedEntry
An occupied entry that holds the value.

VacantEntry
A vacant entry with previous and next vacant indices.

Enums

Entry
An entry within the LazyHashMap.

Type Definitions

EntryMap
The map for the contract storage entries.