Struct ink_storage::lazy::lazy_hmap::OccupiedEntry

pub struct OccupiedEntry<'a, K, V> where
    K: Clone,  { /* fields omitted */ }

An occupied entry that holds the value.

Implementations

impl<'a, K, V> OccupiedEntry<'a, K, V> where
K: Ord + Clone + PackedLayout,
V: PackedLayout,

pub fn key(&self) -> &K

Gets a reference to the key in the entry.

pub fn remove_entry(self) -> (K, V)

Take the ownership of the key and value from the map.

pub fn get(&self) -> &V

Gets a reference to the value in the entry.

pub fn get_mut(&mut self) -> &mut V

Gets a mutable reference to the value in the entry.

If you need a reference to the OccupiedEntry which may outlive the destruction of the Entry value, see into_mut.

pub fn insert(&mut self, new_value: V) -> V

Sets the value of the entry, and returns the entry’s old value.

pub fn remove(self) -> V

Takes the value out of the entry, and returns it.

pub fn into_mut(self) -> &'a mut V

Converts the OccupiedEntry into a mutable reference to the value in the entry with a lifetime bound to the map itself.