Struct ink_storage::lazy::LazyArray

pub struct LazyArray<T, const N: usize> { /* fields omitted */ }

A lazy storage array that spans over N storage cells.

Storage data structure to emulate storage arrays: [T; N].

Note

Computes operations on the underlying N storage cells in a lazy fashion. Due to the size constraints the LazyArray is generally more efficient than the LazyMap for most use cases with limited elements.

This is mainly used as low-level storage primitives by other high-level storage primitives in order to manage the contract storage for a whole chunk of storage cells.

Implementations

impl<T, const N: usize> LazyArray<T, N> where
T: PackedLayout,

pub fn clear_packed_at(&self, index: u32)

Clears the underlying storage of the entry at the given index.

Safety

For performance reasons this does not synchronize the lazy array’s memory-side cache which invalidates future accesses the cleared entry. Care should be taken when using this API.

The general use of this API is to streamline Drop implementations of high-level abstractions that build upon this low-level data structure.

impl<T, const N: usize> LazyArray<T, N>

pub fn new() -> Self

Creates a new empty lazy array.

Note

A lazy array created this way cannot be used to load from the contract storage. All operations that directly or indirectly load from storage will panic.

pub fn capacity(&self) -> u32

Returns the constant capacity of the lazy array.

pub fn key(&self) -> Option<&Key>

Returns the offset key of the lazy array if any.

pub fn put(&mut self, at: u32, new_value: Option<T>)

Puts a new value into the given indexed slot.

Note

Use LazyArray::put_get(None) to remove an element.

impl<T, const N: usize> LazyArray<T, N>

pub fn key_at(&self, at: u32) -> Option<Key>

Returns the offset key for the given index if not out of bounds.

impl<T, const N: usize> LazyArray<T, N> where
T: PackedLayout,

pub fn get(&self, at: u32) -> Option<&T>

Returns a shared reference to the element at the given index if any.

Note

This operation eventually loads from contract storage.

Panics

If the given index is out of bounds.

pub fn get_mut(&mut self, at: u32) -> Option<&mut T>

Returns an exclusive reference to the element at the given index if any.

Note

This operation eventually loads from contract storage.

Panics

If the given index is out of bounds.

pub fn put_get(&mut self, at: u32, new_value: Option<T>) -> Option<T>

Puts the new value into the indexed slot and returns the old value if any.

Note

This operation eventually loads from contract storage.

Prefer LazyArray::put if you are not interested in the old value.

Use LazyArray::put_get(None) to remove an element.

Panics

If the given index is out of bounds.

pub fn swap(&mut self, a: u32, b: u32)

Swaps the values at indices x and y.

Note

This operation eventually loads from contract storage.

Panics

If any of the given indices is out of bounds.