Struct ink_storage::lazy::LazyHashMap

pub struct LazyHashMap<K, V, H> { /* fields omitted */ }

A lazy storage mapping that stores entries under their SCALE encoded key hashes.

Note

This is mainly used as low-level storage primitives by other high-level storage primitives in order to manage the contract storage for a whole mapping of storage cells.

This storage data structure might store its entries anywhere in the contract storage. It is the users responsibility to keep track of the entries if it is necessary to do so.

Implementations

impl<K, V, H> LazyHashMap<K, V, H> where
K: Ord,

pub fn new() -> Self

Creates a new empty lazy hash map.

Note

A lazy map created this way cannot be used to load from the contract storage. All operations that directly or indirectly load from storage will panic.

pub fn key(&self) -> Option<&Key>

Returns the offset key of the lazy map if any.

pub fn put(&mut self, key: K, new_value: Option<V>)

Puts the new value under the given key.

Note

Use LazyHashMap::put(None) in order to remove an element.

Prefer this method over LazyHashMap::put_get if you are not interested in the old value of the same cell index.

Panics

If the lazy hash map is in an invalid state that forbids interaction with the underlying contract storage.

If the decoding of the old element at the given index failed.

impl<K, V, H> LazyHashMap<K, V, H> where
K: Clone + Ord + PackedLayout,
V: PackedLayout,
H: CryptoHash,
Key: From<<H as HashOutput>::Type>,

pub fn entry(&mut self, key: K) -> Entry<'_, K, V>

Gets the given key’s corresponding entry in the map for in-place manipulation.

impl<K, V, H> LazyHashMap<K, V, H> where
K: Ord + Eq + Encode,
V: PackedLayout,
H: CryptoHash,
Key: From<<H as HashOutput>::Type>,

pub fn clear_packed_at<Q>(&self, index: &Q) where
K: Borrow<Q>,
V: PackedLayout,
Q: Ord + Encode + ToOwned<Owned = K>,

Clears the underlying storage of the entry at the given index.

Safety

For performance reasons this does not synchronize the lazy index map’s memory-side cache which invalidates future accesses the cleared entry. Care should be taken when using this API.

The general use of this API is to streamline Drop implementations of high-level abstractions that build upon this low-level data structure.

pub fn get<Q>(&self, index: &Q) -> Option<&V> where
K: Borrow<Q>,
Q: Ord + Encode + ToOwned<Owned = K>,

Returns a shared reference to the value associated with the given key if any.

Panics

If the lazy chunk is in an invalid state that forbids interaction.

If the decoding of the element at the given index failed.

pub fn get_mut<Q>(&mut self, index: &Q) -> Option<&mut V> where
K: Borrow<Q>,
Q: Ord + Encode + ToOwned<Owned = K>,

Returns an exclusive reference to the value associated with the given key if any.

Panics

If the lazy chunk is in an invalid state that forbids interaction.

If the decoding of the element at the given index failed.

pub fn put_get<Q>(&mut self, key: &Q, new_value: Option<V>) -> Option<V> where
K: Borrow<Q>,
Q: Ord + Encode + ToOwned<Owned = K>,

Puts the new value under the given key and returns the old value if any.

Note

Use LazyHashMap::put_get(None) in order to remove an element and retrieve the old element back.

Panics

If the lazy hashmap is in an invalid state that forbids interaction.

If the decoding of the old element at the given index failed.

pub fn swap<Q1, Q2>(&mut self, x: &Q1, y: &Q2) where
K: Borrow<Q1> + Borrow<Q2>,
Q1: Ord + PartialEq<Q2> + Encode + ToOwned<Owned = K>,
Q2: Ord + PartialEq<Q1> + Encode + ToOwned<Owned = K>,

Swaps the values at entries with associated keys x and y.

This operation tries to be as efficient as possible and reuse allocations.

Panics

If the lazy hashmap is in an invalid state that forbids interaction.

If the decoding of one of the elements failed.