Struct ink_storage::lazy::LazyIndexMap

pub struct LazyIndexMap<V> { /* fields omitted */ }

A lazy storage chunk that spans over a whole chunk of storage cells.

Note

This is mainly used as low-level storage primitives by other high-level storage primitives in order to manage the contract storage for a whole chunk of storage cells.

A chunk of storage cells is a contiguous range of 2^32 storage cells.

Implementations

impl<V> LazyIndexMap<V>

pub fn new() -> Self

Creates a new empty lazy map.

Note

A lazy map created this way cannot be used to load from the contract storage. All operations that directly or indirectly load from storage will panic.

pub fn key(&self) -> Option<&Key>

Returns the offset key of the lazy map if any.

pub fn put(&mut self, index: u32, new_value: Option<V>)

Puts the new value at the given index.

Note

Use LazyIndexMap::put(None) in order to remove an element.

Prefer this method over LazyIndexMap::put_get if you are not interested in the old value of the same cell index.

Panics

If the lazy chunk is in an invalid state that forbids interaction.

If the decoding of the old element at the given index failed.

impl<V> LazyIndexMap<V> where
V: PackedLayout,

pub fn clear_packed_at(&self, index: u32)

Clears the underlying storage of the entry at the given index.

Safety

For performance reasons this does not synchronize the lazy index map’s memory-side cache which invalidates future accesses the cleared entry. Care should be taken when using this API.

The general use of this API is to streamline Drop implementations of high-level abstractions that build upon this low-level data structure.

impl<V> LazyIndexMap<V> where
V: PackedLayout,

pub fn key_at(&self, index: u32) -> Option<Key>

Returns an offset key for the given index.

pub fn get(&self, index: u32) -> Option<&V>

Returns a shared reference to the element at the given index if any.

Panics

If the lazy chunk is in an invalid state that forbids interaction.

If the decoding of the element at the given index failed.

pub fn get_mut(&mut self, index: u32) -> Option<&mut V>

Returns an exclusive reference to the element at the given index if any.

Panics

If the lazy chunk is in an invalid state that forbids interaction.

If the decoding of the element at the given index failed.

pub fn put_get(&mut self, index: u32, new_value: Option<V>) -> Option<V>

Puts the new value at the given index and returns the old value if any.

Note

Use LazyIndexMap::put_get(None) in order to remove an element and retrieve the old element back.

Panics

If the lazy chunk is in an invalid state that forbids interaction.

If the decoding of the old element at the given index failed.

pub fn swap(&mut self, x: u32, y: u32)

Swaps the values at indices x and y.

This operation tries to be as efficient as possible and reuse allocations.

Panics

If the lazy chunk is in an invalid state that forbids interaction.

If the decoding of one of the elements failed.