Struct ink_storage::lazy::Mapping

pub struct Mapping<K, V> { /* fields omitted */ }

A mapping of key-value pairs directly into contract storage.

Implementations

impl<K, V> Mapping<K, V> where
K: PackedLayout,
V: PackedLayout,

pub fn insert<Q, R>(&mut self, key: Q, value: &R) where
Q: EncodeLike<K>,
R: EncodeLike<V> + PackedLayout,

Insert the given value to the contract storage.

pub fn get<Q>(&self, key: Q) -> Option<V> where
Q: EncodeLike<K>,

Get the value at key from the contract storage.

Returns None if no value exists at the given key.

pub fn remove<Q>(&self, key: Q) where
Q: EncodeLike<K>,

Clears the value at key from storage.