Struct ink_storage::Box

pub struct Box<T> where
    T: SpreadLayout,  { /* private fields */ }

A dynamically allocated storage entity.

Users can use this in order to make certain SpreadLayout storage entities used in contexts that require a PackedLayout storage entity by simply packing the storage entity within a storage::Box.

Dynamic allocations caused by the creation of storage::Box instances do have some limited overhead:

The dynamic allocation itself has to be provided by some dynamic storage allocator that needs to be invoked.

Each dynamic storage allocation implies roughly 1.12 bits of overhead.

Upon ever first dereferencing of a storage::Box instance a cryptographic hash routine is run in order to compute the underlying storage key.

Use this abstraction with caution due to the aforementioned performance implications.

Implementations

impl<T> Box<T> where
T: SpreadLayout,

pub fn new(value: T) -> Self

Creates a new boxed entity.

impl<T> Box<T> where
T: SpreadLayout,

pub fn get(boxed: &Self) -> &T

Returns a shared reference to the boxed value.

Note

This loads the value from the pointed to contract storage if this did not happen before.

Panics

If loading from contract storage failed.

pub fn get_mut(boxed: &mut Self) -> &mut T

Returns an exclusive reference to the boxed value.

Note

This loads the value from the pointed to contract storage if this did not happen before.

Panics

If loading from contract storage failed.