Module ink_storage::traits

Traits and interfaces to operate with storage entities.

Generally a type is said to be a storage entity if it implements the ink_primitives::traits::Storable trait. This defines certain constants and routines in order to tell a smart contract how to load and store instances of this type from and to the contract’s storage.

The Packed shows that the type can be stored into single storage cell. In most cases, collections(Vec, HashMap, HashSet etc.) work only with packed structures.

If at least one of the type’s fields occupies its own separate storage cell, it is a non-Packed type because it occupies more than one storage cell.

Structs

AutoKey

Auto key type means that the storage key should be calculated automatically.

ManualKey

Manual key type specifies the storage key.

ResolverKey

Resolver key type selects between preferred key and autogenerated key. If the L type is AutoKey it returns auto-generated R else L.

Traits

AutoStorableHint

Automatically returns the type that should be used for storing the value.

LayoutCryptoHasher

Types implementing this trait are supported layouting crypto hashers.

Packed

Trait for describing types that can be read and written to storage while all fields occupy only a single storage cell.

StorableHint

Describes the type that should be used for storing the value and preferred storage key.

StorageKey

Holds storage key for the type.

StorageLayout

Implemented by types that have a storage layout.

Derive Macros

StorableHint

Derives ink_storage’s StorableHint trait for the given struct or enum.

StorageKey

Derives ink_storage’s StorageKey trait for the given struct or enum.

StorageLayout

Derives ink_storage’s StorageLayout trait for the given struct or enum.