Function ink_storage::traits::clear_spread_root

pub fn clear_spread_root<T>(entity: &T, root_key: &Key) where
    T: SpreadLayout, 

Clears the entity from the contract storage using spread layout.

The root key denotes the offset into the contract storage where the instance of type T is being cleared from.

Note

The routine assumes that the instance has previously been stored to the contract storage using spread layout.

Users should prefer using this function directly instead of using the trait methods on SpreadLayout.