Function ink_storage::traits::forward_clear_packed

pub fn forward_clear_packed<T>(entity: &T, ptr: &mut KeyPtr) where
    T: PackedLayout, 

Clears an instance of type T in packed fashion from the contract storage.

Clears the instance from the storage location identified by ptr. The cleared storage entity is expected to be encoded in its packed form.

Note

Use this utility function to use a packed clear operation for the type instead of a spread storage layout clear operation.