Trait ink_storage::traits::PackedAllocate

pub trait PackedAllocate: SpreadAllocate + PackedLayout {
    fn allocate_packed(&mut self, at: &Key);
}

Types that can be default initialized to a single storage cell.

Required methods

fn allocate_packed(&mut self, at: &Key)

Indicates to self that is has just been allocated to the storage.

Note

Most types will have to implement a trivial forwarding to their fields.