Crate ink_storage

The ink_storage utilities used to manipulate and organize contract storage.

Mainly provides entities to work on a contract’s storage as well as high-level collections on top of those. Also provides environmental utilities, such as storage allocators, FFI to interface with FRAME contracts and a primitive blockchain emulator for simple off-chain testing.

Re-exports

pub use ink_storage_traits as traits;

Structs

Lazy

A simple wrapper around a type to store it in a separate storage cell under its own storage key. If you want to update the value, first you need to get it, update the value, and then call set with the new value.

Mapping

A mapping of key-value pairs directly into contract storage.