Crate ink_storage

The ink_storage utilities used to manipulate and organize contract storage.

Mainly provides entities to work on a contract's storage as well as high-level collections on top of those. Also provides environmental utilities, such as storage allocators, FFI to interface with SRML contracts and a primitive blockchain emulator for simple off-chain testing.

Modules

alloc
The default dynamic storage allocator.

collections
High-level collections used to manage storage entities in the persisted contract storage.

lazy
Low-level collections and data structures to manage storage entities in the persisted contract storage.

traits
Traits and interfaces to operate with storage entities.

Structs

Box
A dynamically allocated storage entity.

Lazy
A lazy storage entity.

Memory
An instance that is solely stored within the contract's memory.

Pack
Packs the inner T so that it only occupies a single contract storage cell.

Vec
A contiguous growable array type, written Vec<T> but pronounced 'vector'.