Module ink_storage::collections::hashmap

A storage hash map that allows to associate keys with values.

Structs

HashMap
A hash map operating on the contract storage.

Iter
An iterator over shared references to the elements of a storage hash map.

IterMut
An iterator over shared references to the elements of a storage hash map.

Keys
An iterator over references to the keys of a storage hash map.

OccupiedEntry
An occupied entry that holds the value.

VacantEntry
A vacant entry with previous and next vacant indices.

Values
An iterator over shared references to the values of a storage hash map.

ValuesMut
An iterator over exclusive references to the values of a storage hash map.

Enums

Entry
An entry within the stash.