Module ink_storage::lazy

Low-level collections and data structures to manage storage entities in the persisted contract storage.

Users should generally avoid using these collections directly in their contracts and should instead adhere to the high-level collections found in collections. The low-level collections are mainly used as building blocks for internals of other higher-level storage collections.

These low-level collections are not aware of the elements they manage thus extra care has to be taken when operating directly on them.

Modules

lazy_hmap

A lazy storage mapping that stores entries under their SCALE encoded key hashes.

mapping

A simple mapping to contract storage.

Structs

Lazy

A lazy storage entity.

LazyArray

A lazy storage array that spans over N storage cells.

LazyCell

A lazy storage entity.

LazyHashMap

A lazy storage mapping that stores entries under their SCALE encoded key hashes.

LazyIndexMap

A lazy storage chunk that spans over a whole chunk of storage cells.

Mapping

A mapping of key-value pairs directly into contract storage.