Struct ink_storage::lazy::LazyCell

pub struct LazyCell<T> where
    T: SpreadLayout,  { /* fields omitted */ }

A lazy storage entity.

This loads its value from storage upon first use.

Note

Use this if the storage field does not need to be loaded in some or most cases.

Implementations

impl<T> LazyCell<T> where
T: SpreadLayout,

pub fn new(value: Option<T>) -> Self

Creates an already populated lazy storage cell.

Note

Since this already has a value it will never actually load from the contract storage.

pub fn lazy(key: Key) -> Self

Creates a lazy storage cell for the given key.

Note

This will actually lazily load from the associated storage cell upon access.

impl<T> LazyCell<T> where
T: SpreadLayout,

pub fn get(&self) -> Option<&T>

Returns a shared reference to the value.

Note

This eventually lazily loads the value from the contract storage.

Panics

If decoding the loaded value to T failed.

pub fn get_mut(&mut self) -> Option<&mut T>

Returns an exclusive reference to the value.

Note

This eventually lazily loads the value from the contract storage.

Panics

If decoding the loaded value to T failed.

pub fn set(&mut self, new_value: T)

Sets the value in this cell to value, without executing any reads.

Note

No reads from contract storage will be executed.

This method should be preferred over dereferencing or get_mut in case the returned value is of no interest to the caller.

Panics

If accessing the inner value fails.