Struct ink_storage::Mapping

pub struct Mapping<K, V: Packed, KeyType: StorageKey = AutoKey> { /* private fields */ }

A mapping of key-value pairs directly into contract storage.

Important

The mapping requires its own pre-defined storage key where to store values. By default, the is automatically calculated using AutoKey during compilation. However, anyone can specify a storage key using ManualKey. Specifying the storage key can be helpful for upgradeable contracts or you want to be resistant to future changes of storage key calculation strategy.

This is an example of how you can do this:

use ink::storage::{
    traits::ManualKey,
    Mapping,
};

#[ink(storage)]
#[derive(Default)]
pub struct MyContract {
    balances: Mapping<AccountId, Balance, ManualKey<123>>,
}

impl MyContract {
    #[ink(constructor)]
    pub fn new() -> Self {
        let mut instance = Self::default();
        let caller = Self::env().caller();
        let value: Balance = Default::default();
        instance.balances.insert(&caller, &value);
        instance
    }

}

More usage examples can be found in the ink! examples.

Implementations

impl<K, V, KeyType> Mapping<K, V, KeyType>where V: Packed, KeyType: StorageKey,

pub const fn new() -> Self

Creates a new empty Mapping.

impl<K, V, KeyType> Mapping<K, V, KeyType>where K: Encode, V: Packed, KeyType: StorageKey,

pub fn insert<Q, R>(&mut self, key: Q, value: &R) -> Option<u32>where Q: EncodeLike<K>, R: Storable + EncodeLike<V>,

Insert the given value to the contract storage.

Returns the size in bytes of the pre-existing value at the specified key if any.

Panics

Traps if the encoded value doesn’t fit into the static buffer.

pub fn try_insert<Q, R>(&mut self, key: Q, value: &R) -> Result<Option<u32>>where Q: EncodeLike<K>, R: Storable + EncodeLike<V>,

Try to insert the given value into the mapping under given key.

Fails if value exceeds the static buffer size.

Returns:

Ok(Some(_)) if the value was inserted successfully, containing the size in bytes of the pre-existing value at the specified key if any.

Ok(None) if the insert was successful but there was no pre-existing value.

Err(_) if the encoded value exceeds the static buffer size.

pub fn get<Q>(&self, key: Q) -> Option<V>where Q: EncodeLike<K>,

Get the value at key from the contract storage.

Returns None if no value exists at the given key.

Panics

Traps if the encoded value doesn’t fit into the static buffer.

pub fn try_get<Q>(&self, key: Q) -> Option<Result<V>>where Q: EncodeLike<K>,

Try to get the value at the given key.

Returns:

Some(Ok(_)) containing the value if it existed and was decoded successfully.

Some(Err(_)) if the value existed but its length exceeds the static buffer size.

None if there was no value under this mapping key.

pub fn take<Q>(&self, key: Q) -> Option<V>where Q: EncodeLike<K>,

Removes the value at key, returning the previous value at key from storage.

Returns None if no value exists at the given key.

Panics

Traps if the encoded value doesn’t fit into the static buffer.

Warning

This method uses the unstable interface, which is unsafe and normally is not available on production chains.

pub fn try_take<Q>(&self, key: Q) -> Option<Result<V>>where Q: EncodeLike<K>,

Try to take the value at the given key. On success, this operation will remove the value from the mapping

Returns:

Some(Ok(_)) containing the value if it existed and was decoded successfully.

Some(Err(_)) if the value existed but its length exceeds the static buffer size.

None if there was no value under this mapping key.

Warning

This method uses the unstable interface, which is unsafe and normally is not available on production chains.

pub fn size<Q>(&self, key: Q) -> Option<u32>where Q: EncodeLike<K>,

Get the size in bytes of a value stored at key in the contract storage.

Returns None if no value exists at the given key.

pub fn contains<Q>(&self, key: Q) -> boolwhere Q: EncodeLike<K>,

Checks if a value is stored at the given key in the contract storage.

Returns false if no value exists at the given key.

pub fn remove<Q>(&self, key: Q)where Q: EncodeLike<K>,

Clears the value at key from storage.