Function ink_storage::traits::push_packed_root

pub fn push_packed_root<T>(entity: &T, root_key: &Key) where
    T: PackedLayout, 

Pushes the entity to the contract storage using packed layout.

The root key denotes the offset into the contract storage where the instance of type T is being pushed to.

Note

The routine will push the given entity to the contract storage using packed layout.

Users should prefer using this function directly instead of using the trait methods on PackedLayout.